30.06 update

+++ mainwindow.cpp
@@ -26,7 +26,13 @@ MainWindow::MainWindow(QWidget *parent)
     //context menu for result list
     ui->ResultList->setContextMenuPolicy(Qt::CustomContextMenu);
 
-    //connections to manipulate workers from main thread
+    Ww=new WarningWindow(this);
+
+    //connections between Main Window and Warning Window
+    connect (Ww, &WarningWindow::YesNoClicked, this, &MainWindow::deleteFile);
+    connect (Ww, &WarningWindow::rejected, this, &MainWindow::WWindowClosed);
+
+    //connections for workers manipulation from main thread
     connect (this, &MainWindow::StartHashing1, worker1, &Worker::HashWorker);
     connect (worker1, &Worker::SendItem, this, &MainWindow::GetData);
     connect (this, &MainWindow::StartHashing2, worker2, &Worker::HashWorker);
@@ -58,10 +64,10 @@ MainWindow::~MainWindow()
     delete ui;
 }
 
+/*double click on two dots to view the upper directory,
+on one dot to view the root and on any directory to view what it contains*/
 void MainWindow::on_UpperList_doubleClicked(const QModelIndex &index)
 {
-    /*double click on two dots to view the upper directory,
-    on one dot to view the root and on any directory to view what it contains*/
     ui->UpperList =(QListView*) sender();
     QFileInfo fileinfo =  model->fileInfo(index);
     if (fileinfo.fileName() == "..")
@@ -81,9 +87,9 @@ void MainWindow::on_UpperList_doubleClicked(const QModelIndex &index)
     ui->UpperLine->setText(model->filePath(ui->UpperList->rootIndex()));
 }
 
+//same for the lower list
 void MainWindow::on_LowerList_doubleClicked(const QModelIndex &index)
 {
-    //same for the lower list
     ui->LowerList =(QListView*) sender();
     QFileInfo fileinfo =  model->fileInfo(index);
     if (fileinfo.fileName() == "..")
@@ -119,10 +125,10 @@ void MainWindow::on_CheckButton_clicked()
     }
 }
 
+/*check if selected folders are the same or if one include anothother. Then start the threads.
+ due to this check we can exlude one folder so the threads won't scan it twice */
 bool MainWindow::IsDirNotDifferent(const QString &str1, const QString &str2)
 {
-    /*check if selected folders are the same or if one include anothother. Then start the threads.
-     due to this check we can exlude one folder so the threads won't scan it twice */
     if (str1 == str2)
     {
         ui->HelperLabel->setText("What's the point to select identical directories?");
@@ -167,10 +173,9 @@ void MainWindow::on_ResultList_itemDoubleClicked(QListWidgetItem *item)
     QDesktopServices::openUrl(QUrl("file:///" + FIleInfo.path()));
 }
 
+/*Get data from one of the workers. Create item from it and add it to the Result list*/
 void MainWindow::GetData(const QString &itemPath, const QString &name)
 {
-    /*Get data from one of the workers. Create item from it and add it to the Result list*/
-
     QListWidgetItem *item = new QListWidgetItem(ui->ResultList);
     item->setText("(" + name + ") " + "Path: " + itemPath );
     QFileInfo finfo(itemPath);
@@ -185,10 +190,9 @@ void MainWindow::on_SortButton_clicked()
     ui->ResultList->sortItems(Qt::AscendingOrder);
 }
 
+/*helps to find out both workers finished their tasks*/
 void MainWindow::WorkerFinished()
 {
-    /*helps to find out both workers finished their tasks*/
-
     ++ThreadCounter;
     if (ThreadCounter == NumberOfThreads)
     {
@@ -202,9 +206,9 @@ void MainWindow::WorkerFinished()
     }
 }
 
+/*quit the threads right way but takes long...*/
 void MainWindow::on_StopButton_clicked()
 {
-    /*quit the threads right way but takes long...*/
     ui->StopButton->setEnabled(false);
     ui->HelperLabel->setText("Exiting...Please wait...");
     Mydelay();
@@ -218,6 +222,7 @@ void MainWindow::on_StopButton_clicked()
     spinner->stop();
     ui->HelperLabel->setText("Double click on file to open its directory in explorer");
 }
+
 void MainWindow::on_UpperInsideDirs_toggled(bool checked)
 {
     InsidersAllowedFlag1=checked;
@@ -228,7 +233,7 @@ void MainWindow::on_LowerInsideDirs_toggled(bool checked)
     InsidersAllowedFlag2=checked;
 }
 
-
+/*Synchronize file system lists with line edits */
 void MainWindow::on_UpperLine_textChanged(const QString &arg1)
 {
     if (model->index(arg1).isValid())
@@ -252,18 +257,54 @@ void MainWindow::on_ResultList_customContextMenuRequested(const QPoint &pos)
     Delete.setText("REMOVE");
     Delete.setStatusTip("COMPLETELY remove file from your computer");
     Delete.setIcon(QIcon(":/icons/Stop.png"));
-    connect (&Delete, &QAction::triggered, this, &MainWindow::deleteFile);
+    connect (&Delete, &QAction::triggered, this, &MainWindow::DeleteEnabledChecker);
     menu.addAction(&Delete);
     menu.exec(ui->ResultList->viewport()->mapToGlobal(pos));
     }
 }
 
-//Delete file signal
-void MainWindow::deleteFile()
+/* delete file slot. It can be called from signal by Warning window or by
+ * 'DeleteEnabledChecker' if DeleteEnabled flag is true
+ * (which means user set checked 'Don't ask me again' on Warning window)*/
+void MainWindow::deleteFile(const bool &DeleteEn, const bool& YesOrNo)
 {
+    if (DeleteEn)
+        DeleteEnabled=true;
+    this->setEnabled(true);
+    if (YesOrNo)
+    {
     QListWidgetItem* item = ui->ResultList->currentItem();
     QString path = item->text().remove(0,item->text().indexOf(":") + 2);
     QFile file (path);
     file.remove();
     ui->ResultList->takeItem(ui->ResultList->currentRow());
+    }
+}
+
+// delete selected file
+void MainWindow::keyPressEvent(QKeyEvent *event)
+{
+    if(event->key() == Qt::Key_Delete && ui->ResultList->currentItem() != NULL)
+       DeleteEnabledChecker();
+    if (event->key() == Qt::Key_Control && ui->ResultList->currentItem() != NULL)
+            ui->ResultList->currentItem()->setSelected(true);
+}
+
+// check if user really want to delete chosen file
+void MainWindow::DeleteEnabledChecker()
+{
+    if (DeleteEnabled)
+        deleteFile(DeleteEnabled, true);
+    else
+    {
+        Ww->show();
+        this->setDisabled(true);
+        Ww->setEnabled(true);
+    }
+}
+
+// if Warning window was just closed
+void MainWindow::WWindowClosed()
+{
+    this->setEnabled(true);
 }
